axiom-backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
google-generativeai
sse-starlette
python-multipart
python-dotenv
